split encoder and decoder and make encoder reusable and shared among models

+++ od/train.py
@@ -1,9 +1,11 @@
 import os
-import torch.nn.functional as F
+import sys
+sys.path.append('..')
 import torch
-from torch import nn, optim
-from od import anno, detr_dataset, detr_model, detr_model_fusion, match, eval, anchor
-from common.config import train_annotation_file, train_img_od_dict_file, img_size
+from torch import optim
+from od import anno, detr_dataset, detr_model, match, eval, anchor
+from common.config import train_annotation_file, train_img_od_dict_file, img_size, \
+    train_anchor_batch_size, model_save_dir, device_type, model_save_stride
 from od.config import loss_weights
 import focalloss
 import time
@@ -12,17 +14,13 @@ from torchvision.ops import distance_box_iou_loss
 from torch.utils.data import DataLoader
 from torchvision import transforms
 
-device = torch.device("mps")
-# device = torch.device("cpu")
-anchors = torch.tensor(anchor.generate_anchors(variable_stride=False), device=device)
-model = detr_model.DETR(d_enc=384, d_enc_coord_emb=64, d_dec_coord_emb=32, n_enc_head=6, n_dec_head=6,
-                        n_enc_layer=20, n_dec_layer=12, anchors=anchors, exam_diff=True)
-# model = detr_model2.DETR(d=448, d_img_coord_emb=64, d_anchor_coord_emb=32, n_head=7, d_indic=64, n_attn_layer=20,
-#                          anchors=anchors, exam_diff=True)
+device = torch.device(device_type)
+anchors = torch.tensor(anchor.generate_anchors(stride_ratio=.5), device=device)
+model = detr_model.DETR(d_cont=384, d_head=64, d_enc_coord_emb=64, d_dec_coord_emb=32, d_assist=64,
+                        n_enc_layer=20, n_dec_layer=8, anchors=anchors, exam_diff=True)
 model.to(device)
 n_query = len(anchors)
 
-# cls_loss_fun = nn.CrossEntropyLoss(reduction='none')
 optimizer = optim.Adam(model.parameters(), lr=1e-5)
 
 dicts = anno.build_img_dict(train_annotation_file, train_img_od_dict_file, task='od')
@@ -47,20 +45,21 @@ def train(epoch, batch_size, population, num_sample, weight_recover=0.5, gamma=4
             gt_pos_mask = gt_pos_mask.view(B, 1, n_query) * 1
 
             t = time.time()
-            _, cols = match.assign_query(boxes_gt_xyxy, boxes_pred_xyxy, cids_gt, cls_logits_pred, gt_pos_mask, anchors=None)
+            _, cols = match.assign_query(boxes_gt_xyxy, boxes_pred_xyxy, cids_gt, cls_logits_pred, gt_pos_mask, anchors=anchors)
             t_match = time.time() - t
             cols = torch.tensor(np.stack(cols), device=device)
 
             cls_pos_num = torch.sum(gt_pos_mask, dim=-1)
 
-            # matched_pos_gt = torch.where(cols < cls_pos_num)
-            # matched_pos_gt = [(matched_pos_gt[0][i].item(), matched_pos_gt[1][i].item()) for i in
-            #                   range(len(matched_pos_gt[0]))]
-
             gt_matched_indices_batch = torch.arange(B, device=device).view(B, 1). \
                 expand(B, n_query).contiguous().view(B * n_query)
 
             gt_matched_indices_query = cols.view(B * n_query)
+            for bbb in range(B):
+                tmp = -1
+                for c in sorted(cols[bbb].tolist()):
+                    assert c - tmp == 1
+                    tmp = c
 
             query_pos_mask = (cols < cls_pos_num).view(B * n_query) * 1
             n_pos = query_pos_mask.sum()
@@ -82,7 +81,7 @@ def train(epoch, batch_size, population, num_sample, weight_recover=0.5, gamma=4
             box_loss = box_loss * query_pos_mask
             box_loss = box_loss.sum() / (n_pos + 1e-5)
 
-            loss = cls_loss * 10 + box_loss
+            loss = cls_loss * 5 + box_loss
             optimizer.zero_grad()
             t = time.time()
             loss.backward()
@@ -94,38 +93,42 @@ def train(epoch, batch_size, population, num_sample, weight_recover=0.5, gamma=4
                   f'cl {cls_loss.detach().item() * 1000:.3f}|'
                   f'bl {box_loss.detach().item():.3f}|'
                   f'ac {accu:.3f}|rc {recall:.3f}: {n_tp}/{n_pos}|'
-                  f'dif {enc_diff.item():.3f} {logits_diff.item():.3f}|'
+                  f'dif {enc_diff:.3f} {logits_diff:.3f}|'
                   f'tmch {t_match:.3f}|tbp {t_bp:.3f}|'
-                  # f'match {matched_pos_gt}|'
                   f'img {" ".join(img_id)}')
 
 
 if __name__ == '__main__':
-    model_dir = '/Users/zx/Documents/ml/restart/resources'
-    model_files = os.listdir(model_dir)
-    model_files = [f for f in model_files if f.endswith('.pt') and f.startswith('od_detr_anchor')]
-    if len(model_files) == 0:
+    model_files = os.listdir(model_save_dir)
+    od_anchor_model_files = [f for f in model_files if f.endswith('.pt') and f.startswith('od_detr_anchor')]
+    uni_enc_file = f'{model_save_dir}/uni_enc'
+    if len(od_anchor_model_files) == 0:
         model_path_old = None
-        latest_version = -1
+        latest_version = 0
+        enc_model_files = [f for f in model_files if f.endswith('.pt') and f.startswith('uni_enc')]
+        versions = [int(f.split('.')[0].split('_')[-1]) for f in enc_model_files]
+        enc_latest_version = max(versions)
+        enc_model_path = f'{model_save_dir}/uni_enc_{enc_latest_version}.pt'
+        model.encoder.load_state_dict(torch.load(enc_model_path, map_location=device))
     else:
-        versions = [int(f.split('.')[0].split('_')[-1]) for f in model_files]
+        versions = [int(f.split('.')[0].split('_')[-1]) for f in od_anchor_model_files]
         latest_version = max(versions)
-        model_path_old = f'{model_dir}/od_detr_anchor_{latest_version}.pt'
-        saved_state = torch.load(model_path_old)
-        state = model.state_dict()
-        for k, v in saved_state.items():
-            if k in state:
-                state[k] = v
-        model.load_state_dict(state)
-        # model.load_state_dict(saved_state)
+        model_path_old = f'{model_save_dir}/od_detr_anchor_{latest_version}.pt'
+        saved_state = torch.load(model_path_old, map_location=device)
+        model.load_state_dict(saved_state)
+
     for i in range(500):
-        batch = latest_version + 1 + i
-        train(1, batch_size=2, population=1000, num_sample=batch, weight_recover=0, gamma=4)
-        # train(400, batch_size=2, population=2, num_sample=i, weight_recover=0, gamma=4)
+        n_smp = latest_version + 1 + i
+        # ts = time.time()
+        # train(1, batch_size=train_anchor_batch_size, population=1000, num_sample=n_smp, weight_recover=0, gamma=4)
+        # te = time.time()
+        # print(f'----------------------used {te-ts:.3f} secs---------------------------')
+        train(1000, batch_size=2, population=2, num_sample=i, weight_recover=0, gamma=4)
         # train(2, batch_size=2, population=2, num_sample=i, weight_recover=.5, gamma=2)
-        model_path_new = f'{model_dir}/od_detr_anchor_{batch}.pt'
-        torch.save(model.state_dict(), model_path_new)
-        if model_path_old is not None:
-            os.remove(model_path_old)
-        model_path_old = model_path_new
+        if n_smp % model_save_stride == 0:
+            model_path_new = f'{model_save_dir}/od_detr_anchor_{n_smp}.pt'
+            torch.save(model.state_dict(), model_path_new)
+            if model_path_old is not None:
+                os.remove(model_path_old)
+            model_path_old = model_path_new
         # break
